What am I doing.... Client now has constructor; should be created with constructor. Refactored: moved client reader thread routine, client writer thread routine to separate methods from Start method. Client class should be now functional; not tested yet. +Int64MemoryBlock.Bytes(). Simple server now stops client when stopped.

+++ src/h_sillynet/messageReceiver.go
@@ -22,7 +22,7 @@ func (this *MessageReceiver) Write(x byte) {
 }
 
 func (this *MessageReceiver) ready() bool {
-	return (this.expectedSizeDataPosition == SizeOfInt64) && (this.expectedSize == this.memory.Len())
+	return (this.expectedSizeDataPosition == SizeOfInt64) && (this.expectedSize == int64(this.memory.Len()))
 }
 
 func (this *MessageReceiver) clear() {
